+++ buildSrc/src/main/groovy/net/wequick/gradle/AppPlugin.groovy
@@ -18,9 +18,9 @@ package net.wequick.gradle
 import groovy.io.FileType
 import net.wequick.gradle.aapt.Aapt
 import net.wequick.gradle.aapt.SymbolParser
-import net.wequick.gradle.util.DependenciesUtils
 import net.wequick.gradle.util.JNIUtils
 import org.gradle.api.Project
+import org.gradle.api.artifacts.ResolvedDependency
 
 class AppPlugin extends BundlePlugin {
 
@@ -28,7 +28,7 @@ class AppPlugin extends BundlePlugin {
     private static final int UNSET_TYPEID = 99
     private static final int UNSET_ENTRYID = -1
 
-    protected def compileLibs
+    protected Set<Project> mDependentLibProjects
 
     void apply(Project project) {
         super.apply(project)
@@ -60,15 +60,16 @@ class AppPlugin extends BundlePlugin {
 
         project.afterEvaluate {
             // Get all dependencies with gradle script `compile project(':lib.*')'
-            compileLibs = project.configurations.compile.dependencies.findAll {
+            def libs = project.configurations.compile.dependencies.findAll {
                 it.hasProperty('dependencyProject') &&
                         it.dependencyProject.name.startsWith('lib.')
             }
+            mDependentLibProjects = libs.collect { it.dependencyProject }
             if (isBuildingLibs()) {
                 // While building libs, `lib.*' modules are changing to be an application
                 // module and cannot be depended by any other modules. To avoid warnings,
                 // remove the `compile project(':lib.*')' dependencies temporary.
-                project.configurations.compile.dependencies.removeAll(compileLibs)
+                project.configurations.compile.dependencies.removeAll(libs)
             }
         }
 
@@ -99,8 +100,8 @@ class AppPlugin extends BundlePlugin {
         project.dependencies.add('provided', baseJars)
         //  - lib.*
         def libJarNames = []
-        compileLibs.each {
-            libJarNames += getJarName(it.dependencyProject)
+        mDependentLibProjects.each {
+            libJarNames += getJarName(it)
         }
         if (libJarNames.size() > 0) {
             // Collect the jars with absolute file path, fix issue #65
@@ -112,6 +113,7 @@ class AppPlugin extends BundlePlugin {
 
         // Pre-split all the jar dependencies (deep level)
         def compile = project.configurations.compile
+        compile.exclude group: 'com.android.support', module: 'support-annotations'
         rootExt.preLinkJarDir.listFiles().each { file ->
             if (!file.name.endsWith('D.txt')) return
             file.eachLine { line ->
@@ -138,6 +140,7 @@ class AppPlugin extends BundlePlugin {
             }
         }
         if (isBuildingBundle) {
+            // to support data binding in bundle
             // remove data binding dependency
             compile.exclude group: 'com.android.databinding', module: 'adapters'
         }
@@ -150,26 +153,92 @@ class AppPlugin extends BundlePlugin {
         // Fill extensions
         def variantName = variant.name.capitalize()
         def newDexTaskName = 'transformClassesWithDexFor' + variantName
-        def dexTask = project.hasProperty(newDexTaskName) ? project[newDexTaskName] : variant.dex
+        def dexTask = project.hasProperty(newDexTaskName) ? project.tasks[newDexTaskName] : variant.dex
+        File mergerDir = variant.mergeResources.incrementalFolder
+
         small.with {
             javac = variant.javaCompile
             dex = dexTask
-            processManifest = project['process' + variantName + 'Manifest']
+            processManifest = project.tasks['process' + variantName + 'Manifest']
 
-            packagePath = variant.applicationId.replaceAll('\\.', '/')
+            packageName = variant.applicationId
+            packagePath = packageName.replaceAll('\\.', '/')
             classesDir = javac.destinationDir
             bkClassesDir = new File(classesDir.parentFile, "${classesDir.name}~")
 
-            aapt = project['process' + variantName + 'Resources']
+            aapt = project.tasks['process' + variantName + 'Resources']
             apFile = aapt.packageOutputFile
 
-            symbolFile = new File(aapt.textSymbolOutputDir, 'R.txt')
-            rJavaFile = new File(aapt.sourceOutputDir, "${packagePath}/R.java")
+            File symbolDir = aapt.textSymbolOutputDir
+            File sourceDir = aapt.sourceOutputDir
 
-            mergerXml = new File(variant.mergeResources.incrementalFolder, 'merger.xml')
+            symbolFile = new File(symbolDir, 'R.txt')
+            rJavaFile = new File(sourceDir, "${packagePath}/R.java")
+
+            splitRJavaFile = new File(sourceDir.parentFile, "small/${packagePath}/R.java")
+
+            mergerXml = new File(mergerDir, 'merger.xml')
         }
 
         hookVariantTask()
+    }
+
+    /** Collect the vendor aars (has resources) compiling in current bundle */
+    protected void collectVendorAars(Set<Map> outFirstLevelAars,
+                                     Set<Map> outTransitiveAars) {
+        project.configurations.compile.resolvedConfiguration.firstLevelModuleDependencies.each {
+            collectVendorAars(it, outFirstLevelAars, outTransitiveAars)
+        }
+    }
+
+    protected boolean collectVendorAars(ResolvedDependency node,
+                                        Set<Map> outFirstLevelAars,
+                                        Set<Map> outTransitiveAars) {
+        def group = node.moduleGroup,
+            name = node.moduleName,
+            version = node.moduleVersion
+
+        if (group == '' && version == '') {
+            // Ignores the dependency of local aar
+            return false
+        }
+        if (small.splitAars.find { aar -> group == aar.group && name == aar.name } != null) {
+            // Ignores the dependency which has declared in host or lib.*
+            return false
+        }
+        if (small.retainedAars.find { aar -> group == aar.group && name == aar.name } != null) {
+            // Ignores the dependency of normal modules
+            return false
+        }
+
+        def path = "$group/$name/$version"
+        def aar = [path: path, name: node.name]
+        def resDir = new File(small.aarDir, "$path/res")
+        // If the dependency has resources, collect it
+        if (resDir.exists() && resDir.list().size() > 0) {
+            if (outFirstLevelAars != null && !outFirstLevelAars.contains(aar)) {
+                outFirstLevelAars.add(aar)
+            }
+            if (!outTransitiveAars.contains(aar)) {
+                outTransitiveAars.add(aar)
+            }
+            node.children.each { next ->
+                collectVendorAars(next, null, outTransitiveAars)
+            }
+            return true
+        }
+
+        // Otherwise, check it's children for recursively collecting
+        boolean flag = false
+        node.children.each { next ->
+            flag |= collectVendorAars(next, null, outTransitiveAars)
+        }
+        if (!flag) return false
+
+        if (outFirstLevelAars != null && !outFirstLevelAars.contains(aar)) {
+            outFirstLevelAars.add(aar)
+        }
+        return true
     }
 
     /**
@@ -181,34 +250,16 @@ class AppPlugin extends BundlePlugin {
 
         RootExtension rootExt = project.rootProject.small
 
-        def vendorAars = [] // the vendor aars compiling in current bundle
-        project.configurations.compile.resolvedConfiguration.firstLevelModuleDependencies.each {
-            def group = it.moduleGroup,
-                name = it.moduleName,
-                version = it.moduleVersion
-
-            if (small.splitAars.find { aar -> group == aar.group && name == aar.name } != null) {
-                // Ignore the dependency which has declared in host or lib.*
-                return
-            }
-            if (small.retainedAars.find { aar -> group == aar.group && name == aar.name } != null) {
-                // Ignore the dependency of normal modules
-                return
-            }
-            def resDir = new File(small.aarDir, "$group/$name/$version/res")
-            if (!resDir.exists() || resDir.list().size() == 0) {
-                // Ignored the dependency which does not have any resources
-                return
-            }
-
-            vendorAars.add("$group:$name:$version")
-        }
-        if (vendorAars.size() > 0) {
+        // Check if has any vendor aars
+        def firstLevelVendorAars = [] as Set<Map>
+        def transitiveVendorAars = [] as Set<Map>
+        collectVendorAars(firstLevelVendorAars, transitiveVendorAars)
+        if (firstLevelVendorAars.size() > 0) {
             if (rootExt.strictSplitResources) {
                 def err = new StringBuilder('In strict mode, we do not allow vendor aars, ')
                 err.append('please declare them in host build.gradle:\n')
-                vendorAars.each {
-                    err.append("    - compile('${it}')\n")
+                firstLevelVendorAars.each {
+                    err.append("    - compile('${it.name}')\n")
                 }
                 err.append('or turn off the strict mode in root build.gradle:\n')
                 err.append('    small {\n')
@@ -216,7 +267,8 @@ class AppPlugin extends BundlePlugin {
                 err.append('    }')
                 throw new UnsupportedOperationException(err.toString())
             } else {
-                Log.warn("Using vendor aar(s): ${vendorAars.join('; ')}")
+                def aars = firstLevelVendorAars.collect{ it.name }.join('; ')
+                Log.warn("Using vendor aar(s): $aars")
             }
         }
 
@@ -236,11 +288,11 @@ class AppPlugin extends BundlePlugin {
         def retainedStyleables = []
         def reservedKeys = getReservedResourceKeys()
 
-        bundleEntries.each { k, be ->
+        bundleEntries.each { k, Map be ->
             be._typeId = UNSET_TYPEID // for sort
             be._entryId = UNSET_ENTRYID
 
-            def le = publicEntries.get(k)
+            Map le = publicEntries.get(k)
             if (le != null) {
                 // Use last built id
                 be._typeId = le.typeId
@@ -400,6 +452,7 @@ class AppPlugin extends BundlePlugin {
             // Prepare styleable id maps for resolving R.java
             if (retainedStyleables.size() > 0 && e.typeId == 1) {
                 retainedStyleables.findAll { it.idStrs != null }.each {
+                    // Replace `e.idStr' with `newResIdStr'
                     def index = it.idStrs.indexOf(e.idStr)
                     if (index >= 0) {
                         it.idStrs[index] = newResIdStr
@@ -419,10 +472,104 @@ class AppPlugin extends BundlePlugin {
             it.idStrs = null
         }
 
+        // Collect all the resources for generating a temporary full edition R.java
+        // which required in javac.
+        // TODO: Do this only for the modules who's code really use R.xx of lib.*
+        def allTypes = []
+        def allStyleables = []
+        def addedTypes = [:]
+        libEntries.each { k, e ->
+            if (reservedKeys.contains(k)) return
+
+            if (e.isStyleable) {
+                allStyleables.add(e);
+            } else {
+                if (!addedTypes.containsKey(e.type)) {
+                    // New type
+                    currType = [type: e.vtype, name: e.type, entries: []]
+                    allTypes.add(currType)
+                    addedTypes.put(e.type, currType)
+                } else {
+                    currType = addedTypes[e.type]
+                }
+
+                def entry = [name: e.key, _vs: e.idStr]
+                currType.entries.add(entry)
+            }
+        }
+        retainedTypes.each { t ->
+            def at = addedTypes[t.name]
+            if (at != null) {
+                at.entries.addAll(t.entries)
+            } else {
+                allTypes.add(t)
+            }
+        }
+        allStyleables.addAll(retainedStyleables)
+
+        // Collect vendor types and styleables if needed
+        def vendorEntries = [:]
+        def vendorStyleableKeys = [:]
+        transitiveVendorAars.each { aar ->
+            String path = aar.path
+            File dir = new File(small.aarDir, path)
+            File vendorIdsFile = new File(dir, 'R.txt')
+            def entries = []
+            def styleables = []
+
+            SymbolParser.collectResourceKeys(vendorIdsFile, entries, styleables)
+
+            vendorEntries.put(path, entries)
+            vendorStyleableKeys.put(path, styleables)
+        }
+
+        def vendorTypes = [:]
+        def vendorStyleables = [:]
+        vendorEntries.each { name, es ->
+            allTypes.each { t ->
+                t.entries.each { e ->
+                    def ve = es.find { it.type == t.name && it.name == e.name }
+                    if (ve != null) {
+                        def vendorType
+                        def vts = vendorTypes[name]
+                        if (vts == null) {
+                            vts = vendorTypes[name] = []
+                        } else {
+                            vendorType = vts.find { it.name == t.name }
+                        }
+                        if (vendorType == null) {
+                            vendorType = [:]
+                            vendorType.putAll(t)
+                            vendorType.entries = []
+                            vts.add(vendorType)
+                        }
+                        vendorType.entries.add(e)
+                    }
+                }
+            }
+        }
+        vendorStyleableKeys.each { name, vs ->
+            allStyleables.each { s ->
+                if (vs.contains(s.key)) {
+                    if (vendorStyleables[name] == null) {
+                        vendorStyleables[name] = []
+                    }
+                    vendorStyleables[name].add(s)
+                    return
+                }
+            }
+        }
+
         small.idMaps = staticIdMaps
         small.idStrMaps = staticIdStrMaps
         small.retainedTypes = retainedTypes
         small.retainedStyleables = retainedStyleables
+
+        small.allTypes = allTypes
+        small.allStyleables = allStyleables
+
+        small.vendorTypes = vendorTypes
+        small.vendorStyleables = vendorStyleables
     }
 
     protected int getABIFlag() {
@@ -448,29 +595,37 @@ class AppPlugin extends BundlePlugin {
     }
 
     protected void hookVariantTask() {
+        RootExtension rootExt = project.rootProject.small
+
         // Hook preBuild task to resolve dependent AARs
         project.preBuild.doFirst {
             // Collect dependent AARs
-            RootExtension rootExt = project.rootProject.small
-            def libAars = new HashSet() // the aars compiled in host or lib.*
+            def smallLibAars = new HashSet() // the aars compiled in host or lib.*
             rootExt.preLinkAarDir.listFiles().each { file ->
                 if (!file.name.endsWith('D.txt')) return
+                if (file.name.startsWith(project.name)) return
+
                 file.eachLine { line ->
                     def module = line.split(':')
-                    libAars.add(group: module[0], name: module[1], version: module[2])
+                    if (module.size() == 3) {
+                        smallLibAars.add(group: module[0], name: module[1], version: module[2])
+                    } else {
+                        // If using local aar, the version may be unspecific
+                        smallLibAars.add(group: module[0], name: module[1], version: '')
+                    }
                 }
             }
-            def prjAars = new HashSet() // normal modules who's name does not match Small way - `*.*'
+            def userLibAars = new HashSet() // user modules who's name are not in Small way - `*.*'
             project.rootProject.subprojects {
                 if (it.name.startsWith('lib.')) {
-                    libAars.add(group: it.group, name: it.name, version: it.version)
+                    smallLibAars.add(group: it.group, name: it.name, version: it.version)
                 } else if (it.name != 'app' && it.name != 'small' && it.name.indexOf('.') < 0) {
-                    prjAars.add(group: it.group, name: it.name, version: it.version)
+                    userLibAars.add(group: it.group, name: it.name, version: it.version)
                 }
             }
 
-            small.splitAars = libAars
-            small.retainedAars = prjAars
+            small.splitAars = smallLibAars
+            small.retainedAars = userLibAars
         }
 
         // Hook process-manifest task to remove the `android:icon' and `android:label' attribute
@@ -490,8 +645,8 @@ class AppPlugin extends BundlePlugin {
                     }
                     if (line.indexOf('<application') > 0) {
                         // To support plugin JNI, we overwrite the `android:label' attribute with
-                        // the ABIs flag here. So that at the runtime we can fast extract the
-                        // exact JNIs in the supported ABI. (#87, #79)
+                        // the ABIs flag here. So that at the runtime we can exactly extract the
+                        // usable JNIs in the supported ABI. (#87, #79)
                         int flag = getABIFlag()
                         if (flag != 0) {
                             line += "\n        android:label=\"$flag\""
@@ -532,17 +687,42 @@ class AppPlugin extends BundlePlugin {
 
                 aapt.filterPackage(small.retainedTypes, small.packageId, small.idMaps,
                         small.retainedStyleables)
+
                 Log.success "[${project.name}] slice asset package and reset package id..."
+
+                String pkg = small.packageName
+                // Overwrite the aapt-generated R.java with full edition
+                aapt.generateRJava(small.rJavaFile, pkg, small.allTypes, small.allStyleables)
+                // Also generate a split edition for later re-compiling
+                aapt.generateRJava(small.splitRJavaFile, pkg,
+                        small.retainedTypes, small.retainedStyleables)
+
+                // Overwrite the retained vendor R.java
+                def retainedRFiles = [small.rJavaFile]
+                small.vendorTypes.each { name, types ->
+                    File aarDir = new File(small.aarDir, name)
+                    File manifestFile = new File(aarDir, 'AndroidManifest.xml')
+                    def manifest = new XmlParser().parse(manifestFile)
+                    String aarPkg = manifest.@package
+                    String pkgPath = aarPkg.replaceAll('\\.', '/')
+                    File r = new File(sourceOutputDir, "$pkgPath/R.java")
+                    retainedRFiles.add(r)
+
+                    def styleables = small.vendorStyleables[name]
+                    aapt.generateRJava(r, aarPkg, types, styleables)
+                }
+
+                // Remove unused R.java to fix the reference of shared library resource, issue #63
+                sourceOutputDir.eachFileRecurse(FileType.FILES) { file ->
+                    if (!retainedRFiles.contains(file)) {
+                        file.delete()
+                    }
+                }
+
+                Log.success "[${project.name}] split library R.java files..."
             } else {
                 aapt.resetPackage(small.packageId, small.packageIdStr, small.idMaps)
                 Log.success "[${project.name}] reset resource package id..."
-            }
-
-            // Remove unused R.java to fix the reference of shared library resource, issue #63
-            sourceOutputDir.eachFileRecurse(FileType.FILES) { file ->
-                if (file != rJavaFile) {
-                    file.delete()
-                }
             }
 
             // Repack resources.ap_
@@ -550,27 +730,38 @@ class AppPlugin extends BundlePlugin {
         }
 
         // Hook javac task to split libraries' R.class
+        small.javac.doFirst { t ->
+            // Dynamically provided jars
+            def baseJars = project.fileTree(dir: rootExt.preBaseJarDir, include: ['*.jar'])
+            t.classpath += baseJars
+        }
         small.javac.doLast {
-            File classesDir = it.destinationDir
-            File tempDir = new File(classesDir.parentFile, "${classesDir.name}~")
-            // Filter [package] classes
-            classesDir.renameTo(tempDir)
-            def srcDir = new File(tempDir, small.packagePath)
-            def dstDir = new File(classesDir, small.packagePath)
-            project.copy {
-                from project.fileTree(srcDir)
-                into dstDir
-            }
-            tempDir.deleteDir()
+            if (!small.splitRJavaFile.exists()) return
 
-            Log.success "[${project.name}] split library R.class files..."
+            File classesDir = it.destinationDir
+            File dstDir = new File(classesDir, small.packagePath)
+
+            // Delete the original generated R$xx.class
+            dstDir.listFiles().each { f ->
+                if (f.name.startsWith('R$')) {
+                    f.delete()
+                }
+            }
+            // Re-compile the split R.java to R.class
+            project.ant.javac(srcdir: small.splitRJavaFile.parentFile,
+                    source: it.sourceCompatibility,
+                    target: it.targetCompatibility,
+                    destdir: classesDir)
+
+            Log.success "[${project.name}] split R.class..."
         }
 
         // Hook dex task to split all aar classes.jar
         small.dex.doFirst {
             small.bkAarDir.mkdir()
             small.splitAars.each {
-                String path = "${it.group}/${it.name}/${it.version}"
+                // TODO: Resolve the version conflict
+                String path = "${it.group}/${it.name}" // /${it.version}
                 File dir = new File(small.aarDir, path)
                 if (dir.exists()) {
                     File todir = new File(small.bkAarDir, path)
@@ -578,10 +769,6 @@ class AppPlugin extends BundlePlugin {
                 }
             }
             Log.success "[${project.name}] split aar classes..."
-        }
-        small.dex.doLast {
-            project.ant.move(file: small.bkAarDir, tofile: small.aarDir)
-            small.bkAarDir.delete()
         }
 
         // Hook clean task to unset package id