+++ buildSrc/src/main/groovy/net/wequick/gradle/RootPlugin.groovy
@@ -123,6 +123,7 @@ class RootPlugin extends BasePlugin {
             }
         }
 
+        // todo need to find a better way
         File libs = new File(lib.projectDir, 'libs')
         if (libs.exists()) {
             project.copy {
@@ -135,7 +136,8 @@ class RootPlugin extends BasePlugin {
         ext.explodeAarDirs.each {
             // explodedDir: **/exploded-aar/$group/$artifact/$version
             File version = it
-            File jarFile = new File(version, 'jars/classes.jar')
+            File jarDir = new File(version, 'jars')
+            File jarFile = new File(jarDir, 'classes.jar')
             if (!jarFile.exists()) return
 
             File artifact = version.parentFile
@@ -148,6 +150,21 @@ class RootPlugin extends BasePlugin {
                 from jarFile
                 into preJarDir
                 rename {destFile.name}
+            }
+
+            // Check if exists `jars/libs/*.jar' and copy
+            File libDir = new File(jarDir, 'libs')
+            libDir.listFiles().each { jar ->
+                if (!jar.name.endsWith('.jar')) return
+
+                destFile = new File(preJarDir, "${group.name}-${artifact.name}-${jar.name}")
+                if (destFile.exists()) return
+
+                project.copy {
+                    from jar
+                    into preJarDir
+                    rename {destFile.name}
+                }
             }
         }