+++ buildSrc/CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.7.0 (2016-05-05)
+
+Features:
+
+  - 插件模块支持依赖本地aar模块
+
+Bugfixes:
+
+  - 修正`lib.*`插件在`buildLib`时漏分离`support-annotations.jar`
+  - 修正`lib.*`插件相互依赖时可能出现的类找不到问题 (#117)
+  - 修正_非强制分离_模式下深层次依赖的R.java未被正确处理的问题 (#81, #114 @peacepassion)
+  - 修正在执行`buildBundle`时，`lib.*`插件可能出现的`找不到Small类`错误
+  - 修正在执行`buildBundle`时，`lib.*`插件可能出现的`transformNative_libsWithSyncJniLibsForRelease`任务异常
+
+## 0.6.0 (2016-04-21)
+
+Features:
+
+  - 支持`app.*`在代码中直接引用`lib.*`的资源`R.xx`
+  - 支持不设置宿主签名时可以正常调试运行
+
 ## 0.5.0 (2016-04-12)
 
 Features: